+++ src/hooks/client.ts
@@ -13,7 +13,6 @@ import {
 	getIsArchiveDeviceQueryKey,
 	isArchiveDeviceQueryOptions,
 } from '../lib/react-query/client.js'
-import { getProjectsQueryKey } from '../lib/react-query/projects.js'
 
 /**
  * Access a client API instance. If a ClientApiContext provider is not
@@ -130,52 +129,6 @@ export function useSetOwnDeviceInfoMutation() {
 		onSuccess: () => {
 			queryClient.invalidateQueries({
 				queryKey: getDeviceInfoQueryKey(),
-			})
-		},
-	})
-
-	return {
-		mutate,
-		reset,
-		status,
-	}
-}
-
-export function useCreateProjectMutation() {
-	const queryClient = useQueryClient()
-	const clientApi = useClientApi()
-
-	const { mutate, status, reset } = useMutation({
-		mutationFn: async (opts?: { name?: string; configPath?: string }) => {
-			// Have to avoid passing `undefined` explicitly
-			// See https://github.com/digidem/rpc-reflector/issues/21
-			return opts ? clientApi.createProject(opts) : clientApi.createProject()
-		},
-		onSuccess: () => {
-			queryClient.invalidateQueries({
-				queryKey: getProjectsQueryKey(),
-			})
-		},
-	})
-
-	return {
-		mutate,
-		reset,
-		status,
-	}
-}
-
-export function useLeaveProjectMutation() {
-	const queryClient = useQueryClient()
-	const clientApi = useClientApi()
-
-	const { mutate, status, reset } = useMutation({
-		mutationFn: async (projectPublicId: string) => {
-			return clientApi.leaveProject(projectPublicId)
-		},
-		onSuccess: () => {
-			queryClient.invalidateQueries({
-				queryKey: getProjectsQueryKey(),
 			})
 		},
 	})
